refactor: migrate modules from cjs to mjc
